There are different types of gui properties

+++ src/Raylib.NET.Example/Program.cs
@@ -11,7 +11,7 @@ while (!WindowShouldClose())
 {
     BeginDrawing();
 
-    ClearBackground(GetColor((uint)GuiGetStyle(GuiControl.DEFAULT, GuiDefaultProperty.BACKGROUND_COLOR)));
+    ClearBackground(GetColor((uint)GuiGetStyle(GuiControl.DEFAULT, (int)GuiDefaultProperty.BACKGROUND_COLOR)));
 
     DrawTexture(texture, 10, 40, Color.WHITE);
